+++ packages/xchain-thorchain-amm/src/wallet.ts
@@ -7,7 +7,6 @@ import { Client as CosmosClient } from '@xchainjs/xchain-cosmos'
 import { Client as DogeClient } from '@xchainjs/xchain-doge'
 import { Client as EthClient } from '@xchainjs/xchain-ethereum'
 import { Client as LtcClient } from '@xchainjs/xchain-litecoin'
-import { Client as TerraClient } from '@xchainjs/xchain-terra'
 import { Client as ThorClient, ThorchainClient } from '@xchainjs/xchain-thorchain'
 import { ThorchainQuery } from '@xchainjs/xchain-thorchain-query'
 import { AssetBTC, Chain, assetToString, eqAsset } from '@xchainjs/xchain-util'
@@ -24,7 +23,21 @@ type AllBalances = {
 const chainIds = {
   [Network.Mainnet]: 'thorchain-mainnet-v1',
   [Network.Stagenet]: 'chain-id-stagenet',
-  [Network.Testnet]: 'thorchain-testnet-v2',
+  [Network.Testnet]: 'deprecated',
+}
+const clientUrl = {
+  [Network.Stagenet]: {
+    node: 'https://stagenet-thornode.ninerealms.com',
+    rpc: 'https://stagenet-rpc.ninerealms.com',
+  },
+  [Network.Mainnet]: {
+    node: 'https://thornode.ninerealms.com',
+    rpc: 'https://rpc.ninerealms.com',
+  },
+  [Network.Testnet]: {
+    node: 'deprecated',
+    rpc: 'deprecated',
+  },
 }
 
 /**
@@ -50,9 +63,8 @@ export class Wallet {
       BCH: new BchClient(settings),
       BTC: new BtcClient(settings),
       DOGE: new DogeClient(settings),
-      TERRA: new TerraClient(settings),
       ETH: new EthClient(settings),
-      THOR: new ThorClient({ ...settings, chainIds }),
+      THOR: new ThorClient({ ...settings, chainIds, clientUrl }),
       LTC: new LtcClient(settings),
       BNB: new BnbClient(settings),
       GAIA: new CosmosClient(settings),